ca: remove actingSecondaryCA
This commit removes the actingSecondaryCA field, and removes the stateLock around it. This field
was acting as a proxy for provider != nil, so replace it with that check instead.

This is being done to remove some unnecessary references to c.state, in preparations for extracting
types for primary/secondary.

Also does a small refactor in secondaryUpdateRoots to make the flow more obvious.